Remove old account implementation (#1036)
* Remove old account implementation

* Delete group sidebar

* Remove timetable syncing and groups
