+++ client/src/components/sidebar/UserAccount.tsx
@@ -1,15 +1,12 @@
 import { LoginRounded, LogoutRounded } from '@mui/icons-material';
 import { Button, IconButton, Tooltip } from '@mui/material';
 import { styled } from '@mui/system';
-import React, { useContext, useState } from 'react';
-
+import React, { useState } from 'react';
 import { API_URL } from '../../api/config';
-import { undefinedUser, UserContext } from '../../context/UserContext';
-import { DisplayTimetablesMap } from '../../interfaces/Periods';
 import storage from '../../utils/storage';
 import { createDefaultTimetable } from '../../utils/timetableHelpers';
 import StyledDialog from '../StyledDialog';
-import UserProfile from './groupsSidebar/friends/UserProfile';
+import UserProfile from './friends/UserProfile';
 
 interface UserAccountProps {
   collapsed: boolean;
@@ -45,26 +42,10 @@ const ExpandedContainer = styled('div')`
   padding: 10px 12px;
 `;
 
-export interface User {
-  userID: string;
-  firstname: string;
-  lastname: string;
-  email: string;
-  profileURL: string;
-  createdAt: string;
-  lastLogin: string;
-  loggedIn: boolean;
-  friends: User[];
-  incoming: User[];
-  outgoing: User[];
-  timetables: DisplayTimetablesMap;
-}
-
 const UserAccount: React.FC<UserAccountProps> = ({ collapsed }) => {
   const [windowLocation, setWindowLocation] = useState('');
   const [logoutDialog, setLogoutDialog] = useState(false);
 
-  const { user, setUser } = useContext(UserContext);
   const loginCall = async () => {
     setWindowLocation(window.location.href);
     try {
@@ -72,6 +53,14 @@ const UserAccount: React.FC<UserAccountProps> = ({ collapsed }) => {
     } catch (error) {
       console.log(error);
     }
+  };
+
+  const user = {
+    userID: '',
+    firstname: 'First',
+    lastname: 'Last',
+    email: 'example@user.com',
+    profileURL: '',
   };
 
   const logoutCall = async () => {
@@ -83,7 +72,6 @@ const UserAccount: React.FC<UserAccountProps> = ({ collapsed }) => {
       console.log(error);
     }
     window.location.replace(windowLocation);
-    setUser(undefinedUser);
     storage.set('timetables', createDefaultTimetable(undefined));
   };
   if (!user.userID) {
