+++ client/src/components/sidebar/friends/FriendsTablist.tsx
@@ -5,7 +5,7 @@ import Tab from '@mui/material/Tab';
 import Tabs from '@mui/material/Tabs';
 import * as React from 'react';
 
-import { User } from '../../UserAccount';
+import { User } from './User';
 import AddAFriendTab from './AddAFriendTab';
 import RequestsTab from './RequestsTab';
 import YourFriendsTab from './YourFriendsTab';
