+++ client/src/components/sidebar/friends/RequestsTab.tsx
@@ -3,9 +3,9 @@ import { Check, Close } from '@mui/icons-material';
 import { IconButton, Tooltip } from '@mui/material';
 import React from 'react';
 
-import { API_URL } from '../../../../api/config';
-import NetworkError from '../../../../interfaces/NetworkError';
-import { User } from '../../UserAccount';
+import { API_URL } from '../../../api/config';
+import NetworkError from '../../../interfaces/NetworkError';
+import { User } from './User';
 import UserProfile from './UserProfile';
 
 const StyledFriendsListContainer = styled('div')`
