+++ client/src/components/sidebar/friends/UserProfile.tsx
