+++ client/src/index.tsx
@@ -14,7 +14,6 @@ import EventShareModal from './components/EventShareModal';
 import LandingPage from './components/landingPage/LandingPage';
 import AppContextProvider from './context/AppContext';
 import CourseContextProvider from './context/CourseContext';
-import UserContextProvider from './context/UserContext';
 import * as swRegistration from './serviceWorkerRegistration';
 
 Sentry.init({
@@ -30,24 +29,22 @@ const Root: React.FC = () => {
     <ApolloProvider client={client}>
       <AppContextProvider>
         <CourseContextProvider>
-          <UserContextProvider>
-            <BrowserRouter
-              future={{
-                v7_startTransition: true,
-                v7_relativeSplatPath: true,
-              }}
-            >
-              <Routes>
-                {hasVisited ? (
-                  <Route element={<App />} path="/">
-                    <Route path="/event/:encrypted" element={<EventShareModal />} />
-                  </Route>
-                ) : (
-                  <Route element={<LandingPage />} path="/" />
-                )}
-              </Routes>
-            </BrowserRouter>
-          </UserContextProvider>
+          <BrowserRouter
+            future={{
+              v7_startTransition: true,
+              v7_relativeSplatPath: true,
+            }}
+          >
+            <Routes>
+              {hasVisited ? (
+                <Route element={<App />} path="/">
+                  <Route path="/event/:encrypted" element={<EventShareModal />} />
+                </Route>
+              ) : (
+                <Route element={<LandingPage />} path="/" />
+              )}
+            </Routes>
+          </BrowserRouter>
         </CourseContextProvider>
       </AppContextProvider>
     </ApolloProvider>
